Modification of time coordinates to use float64 and docstring improvement.

+++ lib/improver/tests/ensemble_calibration/ensemble_calibration/helper_functions.py
@@ -113,7 +113,7 @@ def set_up_probability_above_threshold_spot_cube(
     time_origin = "hours since 1970-01-01 00:00:00"
     calendar = "gregorian"
     tunit = Unit(time_origin, calendar)
-    cube.add_dim_coord(DimCoord(np.array([412227], dtype=np.float32),
+    cube.add_dim_coord(DimCoord(np.array([412227], dtype=np.float64),
                                 "time", units=tunit), 1)
     cube.add_dim_coord(
         DimCoord(np.arange(9, dtype=np.float32), long_name='locnum',
@@ -149,9 +149,9 @@ def set_up_probability_above_threshold_spot_temperature_cube():
             data, "air_temperature", "degreesC"))
 
 
-def set_up_cube(data, name, units, realizations=np.array([0, 1, 2],
-                dtype=np.float32), timesteps=1, y_dimension_length=3,
-                x_dimension_length=3):
+def set_up_cube(data, name, units,
+                realizations=np.array([0, 1, 2], dtype=np.float32),
+                timesteps=1, y_dimension_length=3, x_dimension_length=3):
     """Create a cube containing multiple realizations."""
     try:
         cube = Cube(data, standard_name=name, units=units)
@@ -168,7 +168,7 @@ def set_up_cube(data, name, units, realizations=np.array([0, 1, 2],
     num1 = cf_units.date2num(dt1, time_origin, calendar)
     num2 = cf_units.date2num(dt2, time_origin, calendar)
     cube.add_dim_coord(DimCoord(np.linspace(num1, num2, timesteps,
-                                            dtype=np.float32),
+                                            dtype=np.float64),
                                 "time", units=tunit), 1)
     cube.add_dim_coord(DimCoord(np.linspace(-45.0, 45.0, y_dimension_length,
                                             dtype=np.float32),
@@ -198,7 +198,7 @@ def set_up_spot_cube(data, phenomenon_standard_name, phenomenon_units):
     time_origin = "hours since 1970-01-01 00:00:00"
     calendar = "gregorian"
     tunit = Unit(time_origin, calendar)
-    cube.add_dim_coord(DimCoord(np.array(412227, dtype=np.float32),
+    cube.add_dim_coord(DimCoord(np.array(412227, dtype=np.float64),
                                 "time", units=tunit), 1)
     cube.add_dim_coord(DimCoord(np.arange(9, dtype=np.float32),
                                 long_name='locnum',
