+++ lib/improver/tests/utilities/test_load.py
@@ -90,7 +90,7 @@ class Test_load_cube(IrisTest):
         self.cube = set_up_temperature_cube()
         save_netcdf(self.cube, self.filepath)
         self.realization_points = np.array([0, 1, 2], dtype=np.float32)
-        self.time_points = np.array([412227], dtype=np.float32)
+        self.time_points = np.array([412227], dtype=np.float64)
         self.latitude_points = np.array([-45., 0., 45.], dtype=np.float32)
         self.longitude_points = np.array([120., 150., 180.], dtype=np.float32)
 
@@ -259,7 +259,7 @@ class Test_load_cubelist(IrisTest):
         self.cube = set_up_temperature_cube()
         save_netcdf(self.cube, self.filepath)
         self.realization_points = np.array([0, 1, 2])
-        self.time_points = np.array(412227, dtype=np.float32)
+        self.time_points = np.array(412227, dtype=np.float64)
         self.latitude_points = np.array([-45., 0., 45.], dtype=np.float32)
         self.longitude_points = np.array([120., 150., 180.], dtype=np.float32)
         self.low_cloud_filepath = os.path.join(self.directory, "low_cloud.nc")
